Move option and result assertions out of modules
This is a breaking change, but I decide it was better to make it now
than, you know, after people are using it.

In practice, the `Option` and `Result` modules were conflicting with the
Belt.* and Js.* modules of the same name if you used `open Zora`.

I was finding it just created confusion, and decided to nip it in the
bud.